requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.2.1
colorama==0.4.6
flask==3.1.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
pefile==2023.2.7
psutil==7.1.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pywin32-ctypes==0.2.3
pyyaml==6.0.2
setuptools==80.9.0
werkzeug==3.1.3
